MSAT_ELMS/requirements.txt
Django==2.2
django-bootstrap-modal-forms==1.4.1
django-widget-tweaks==1.4.2
selenium==3.14.0
pytz==2018.5